+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 5.2.0-SNAPSHOT
+### 5.2.0
 - Update to Minecraft 1.20.6
 - Fix game crashing/freezing when world name contains illegal characters (#67)
 - Fix "This world contains data from an old version" on new worlds with "Dynamic World Management" enabled (#246)